Add trampolines to diagnostic.
The trampolines are nearby labels that contain long jumps.

We use JB commands that require nearby destinations.  So, we setup
trampolines at nearby locations, and have them jump to (potentially
faraway) handlers.

We are adding so much code to diagnostic.asm that soon the handlers
will be too far away for a short jump (like JB) requires.