Replace per-bit polling with per-byte polling.
When we begin reading data, the SD card will send us start codes that
do not begin with a low bit.  So, our previous strategy of using a low
bit as a signal to start receiving SPI bits will not work.

The new strategy is to ready each byte that the card sends, and check
to see if that byte is a start byte.