Only read as many SD response bytes as required.
Previously, whenever we sent a command, we would read five bytes of
response data back.  Many responses were only one byte long.  In this
case, we read an extra four bytes of data.

In the best case scenario, this would amount to a small waste of time.
We would throw out the data.  In the worst case scenario, we would
have begun to consume a data packet.  This would make it hard to write
code that cleanly requests and receives a data packet.

Now, when we send a command whose response is only one byte, we now
only read one byte of response data.  (We still read and ignore the
extra padding byte.)  So, we will never accidentally begin to consume
a data packet.  Instead, the data-packet-consumption code can exist on
its own, cleanly.