Eliminate the idea of a bootloader_loader.
That is, make the following changes:
  * Replace the term 'bootloader' with 'os'.
  * Replace the term 'bootloader_loader' with 'bootloader'.

In a "normal" PC, the BIOS (or the bootloader loader) loads the
bootloader from the boot sector, which in turn loads the OS.  This
strategy adds flexibility to the system: the BIOS code becomes
independent of the filesystem and the OS.

However, when the system will only be used for one filesystem and one
OS.  And, while writing the BIOS bootloader was fun, writing a second,
boot-sector bootloader will probably be boring.

For these reasons, we eliminate the idea of a boot-sector bootloader
altogether.  The BIOS is the whole bootloader-- it mounts the FAT32
filesystem, reads the file named BOOT.IMG into upper RAM, and jumps
in.  (This functionality was implemented in previous commits.  In the
current commit, I simply acknowledge that the process has become
simpler by changing some names.)